+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ## [unreleased]
 
 ### Added
+- Added 'cached_is_valid' column to 'taxon_names' [#2242]
 - Added `destroyed_redirect` to object radial JSON
 - "Not specified" facet to Filter nomenclature [#2226]
 - New extract task interface [#1934]
@@ -18,6 +19,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Image matrix link in Interactive keys task
 - Export scss vars to javascript
 - Pagination count in Filter nomenclature
+- OTU depictions column on view mode in Image matrix task
+- Grid table component
+- SVG Icon component
+- OTU depictions draggable in image matrix
 
 ### Changed
 - Removed `allow_destroy` from object radial JSON
@@ -28,6 +33,8 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Upgraded to Node 14 LTS
 - Updated Ruby gems and Node packages
 - node-sass to dart-sass
+- Refactor image matrix edit table
+- Webpack configuration to export sass vars
 
 ### Fixed
 - JSON for geographic area parents (no parent raise)
@@ -35,6 +42,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - 404 error when deleting records from data interfaces [#2223]
 - Rank order on New combination preview
 - Redirect after destroy a combination [#2169]
+- Drag and drop depictions in Image Matrix
 
 [#1934]: https://github.com/SpeciesFileGroup/taxonworks/pull/1934
 [#2169]: https://github.com/SpeciesFileGroup/taxonworks/pull/2169